Added giggle_window_show_diff_window and call it from main on --diff
Moved the functionality from the action callback to the new function. Since the main window won't be realized until after the configuration callback is run we will wait with realizing the diff window until after the main window is realized. Otherwise the main window will get the focus as it was the last opened window.

This solves GGL-64.